pyproject.toml
[tool.poetry]
name = "musictobelightapi"
version = "0.1.0"
description = ""
authors = ["catalunha <catalunha.mj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.1.0"
python-decouple = "^3.8"
dj-database-url = "^2.1.0"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-storages = {extras = ["s3"], version = "^1.14.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
